+++ server/api/shopDetails.php
@@ -6,13 +6,11 @@ require_once( API_PATH . '/inc/cors.php');
 
 $connect = connect();
 
-$sql = "SELECT p.id, p.sn, p.datum,
-c.name as city, r.name as region, cn.name as country FROM " . DB_TBL_PIVOFLOW .
+$sql = "SELECT p.sn, c.name as city, r.name as region, cn.name as country, " .
+" AVG(p.temp) as tempAvg, MAX(p.temp) as tempMax, MIN(p.temp) as tempMin FROM " . DB_TBL_PIVOFLOW .
 " p INNER JOIN " . DB_TBL_CITIES . " c ON p.cityId=c.id" .
 " INNER JOIN " . DB_TBL_REGIONS . " r ON c.regionId = r.id" .
 " INNER JOIN " . DB_TBL_COUNTRIES . " cn ON r.countryId = cn.id";
-//simulating distinct only for p.sn to show rows-shops with different SN only
-$sql .= " WHERE p.id IN (SELECT MAX(p2.id) FROM " . DB_TBL_PIVOFLOW . " AS p2 GROUP BY p2.sn)";
 
 if($_SERVER['REQUEST_METHOD'] == 'POST') {
 	//get data from JSON POST
@@ -28,23 +26,18 @@ if($_SERVER['REQUEST_METHOD'] == 'POST') {
 		$sql .= ' AND p.datum <= \'' . $endDate . '\'';
 	}
 	
-	if( isset($data['search']) && (strlen($data['search']) > 0)  ){
-		$search = trim($data['search']);
-		if (strlen($search) > 2) {
-			$sql .= ' AND p.sn LIKE "%' . $search . '%"';
-		}
-	}
 	
-	if( isset($data['cityId']) && (strlen($data['cityId']) > 0)  ){
-		$sql .= ' AND p.cityId = ' . $data['cityId'];
-	} else if ( isset($data['regionId']) && (strlen($data['regionId']) > 0) ) {
-		//handle when cityId is not set, but regionId is set
-	} else if ( isset($data['countryId']) && (strlen($data['countryId']) > 0) ) {
-		//handle when cityId and regionId are not set, but countryId is set
+	if( isset($data['shopIds']) && is_array($data['shopIds']) && !empty($data['shopIds']) ){
+		$sql .= ' AND p.sn IN ("' . implode('","', $data['shopIds']) . '")';
+	} else {
+		//must have narrowed search with shopIds or die with status 400 - bad request
+		http_response_code(400);
+		die();
 	}
 }
-//sort from newest to oldest by last activity
-$sql .= ' ORDER BY datum DESC LIMIT 200';
+
+$sql .= ' GROUP BY p.sn';
+$sql .= ' LIMIT 200';
 
 //echo $sql;
 
@@ -54,18 +47,19 @@ if(!$result = $connect->query($sql)){
 
 while($row = $result->fetch_assoc()){
 	$temp = [
-        "id" => (int) $row['id'],
         "sn" => $row['sn'],
 		"city" => $row['city'],
 		"region" => $row['region'],
 		"country" => $row['country'],
-        "datum" => $row['datum'],
+        "tempAvg" => $row['tempAvg'],
+        "tempMax" => $row['tempMax'],
+        "tempMin" => $row['tempMin'],
     ];
-    $home[] = $temp;
+    $details[] = $temp;
 }
 
 $response = [
-    "data" => $home,
+    "data" => $details,
 ];
 
 $result->free();