+++ src/containers/Home/HomePage.api.js
@@ -3,12 +3,12 @@ import { API_URL } from '../../constants/config';
 
 const HomePageApi = {
 	fetchShops: ({
-			startDate=null,
-			endDate=null,
-			search=null,
-            countryId=null,
-            regionId=null,
-            cityId=null,
+		startDate=null,
+		endDate=null,
+		search=null,
+        countryId=null,
+        regionId=null,
+        cityId=null,
 	}) => {
 
 		let payload = {
@@ -33,6 +33,33 @@ const HomePageApi = {
 			response => response.json()
 		);
 	},
+
+	fetchShopsDetails: ({
+        startDate=null,
+        endDate=null,
+        shopIds = [],
+    }) => {
+
+		let payload = {
+			startDate,
+			endDate,
+			shopIds,
+		};
+
+		let config = {
+			method: 'post',
+			headers: {
+				'Accept': 'application/json',
+				'Content-Type': 'application/json',
+			},
+			body: JSON.stringify(payload),
+		};
+
+		return fetch(API_URL + '/shopDetails', config).then(
+			response => response.json()
+		);
+	},
+
 };
 
 export default HomePageApi;