+++ src/containers/Home/HomePage.js
@@ -9,9 +9,11 @@ import {
 	HOMEPAGE_SET_END_DATE,
 	fetchShops,
 	setSearchValue,
+	setSelectedShops,
+	fetchShopsDetails,
 } from './HomePage.actions';
 
-import { find } from 'lodash';
+import { find, filter } from 'lodash';
 
 import Paper from 'material-ui/Paper';
 import DatePicker from 'material-ui/DatePicker';
@@ -30,6 +32,7 @@ import LineChart from '../../components/Charts/LineChart/LineChart';
 
 import UpdateDataButton from '../../components/UpdateDataButton/UpdateDataButton';
 import ShopsTable from '../../components/ShopsTable/ShopsTable';
+import ShopsDetailsTable from '../../components/ShopsDetailsTable/ShopsDetailsTable';
 import SearchBox from '../../components/SearchBox/SearchBox';
 
 const paperStyle = {
@@ -58,6 +61,16 @@ class HomePage extends React.Component {
 		this.props.setEndDate(date);
 	}
 
+	handleRowSelection (selectedRowIds) {
+		this.props.setSelectedShops(selectedRowIds);
+		let params = {
+			startDate: this.props.datePickerData.startDate,
+			endDate: this.props.datePickerData.endDate,
+			shopIds: selectedRowIds,
+		};
+		this.props.fetchShopsDetails(params);
+	}
+
 	handleGetData () {
 		let params = {
 			startDate: this.props.datePickerData.startDate,
@@ -75,7 +88,6 @@ class HomePage extends React.Component {
 	}
 
 	render() {
-
 		const progressLoader = (<aside className="homePageShopsProgressHolder">
 			<CircularProgress mode="indeterminate" size={100} thickness={10}/>
 		</aside>);
@@ -96,14 +108,32 @@ class HomePage extends React.Component {
 		} else if (this.props.shops.items.length) {
 			dataTable = (
 				<section className="p-">
-					<h3 className="text-left color-blue pb-">Data visualization:</h3>
+					<h3 className="text-left color-blue pb-">Available shops:</h3>
 					<Paper style={paperStyle} zDepth={2}>
-						<ShopsTable data={this.props.shops.items} />
+						<ShopsTable data={this.props.shops.items} onRowSelection={this.handleRowSelection.bind(this)}
+						/>
 					</Paper>
 				</section>
 			);
 		} else if (!this.props.shops.items.length && this.props.shops.lastUpdated) {
 			dataTable = noTableData;
+		}
+
+
+		let selectedShopsDetailsTable;
+		if (this.props.shopsDetails.isFetching) {
+			selectedShopsDetailsTable = progressLoader;
+		} else if (this.props.shopsDetails.items.length) {
+			selectedShopsDetailsTable = (
+				<section className="p-">
+					<h3 className="text-left color-blue pb-">Selected shops details:</h3>
+					<Paper style={paperStyle} zDepth={2}>
+						<ShopsDetailsTable data={this.props.shopsDetails.items} />
+					</Paper>
+				</section>
+			);
+		} else if (!this.props.shopsDetails.items.length && this.props.shopsDetails.lastUpdated) {
+			selectedShopsDetailsTable = noTableData;
 		}
 
 		return (
@@ -152,6 +182,8 @@ class HomePage extends React.Component {
 				</section>
 
 				{dataTable}
+
+				{selectedShopsDetailsTable}
 
 				<section style={{overflow:'hidden', position:'relative'}}>
 					<h3 className="text-left color-blue p- pb0">Data visualization:</h3>
@@ -202,6 +234,8 @@ HomePage.propTypes = {
 	datePickerData: PropTypes.object.isRequired,
 
 	shops: PropTypes.object.isRequired,
+	shopsDetails: PropTypes.object.isRequired,
+	selectedShops: PropTypes.array,
 	search: PropTypes.object,
 	setSearchValue: PropTypes.func,
 
@@ -210,6 +244,8 @@ HomePage.propTypes = {
 	setEndDate: PropTypes.func,
 
 	fetchShops: PropTypes.func,
+	fetchShopsDetails: PropTypes.func,
+	setSelectedShops: PropTypes.func,
 
 	sideMenu: PropTypes.object,
 	selectedCountry: PropTypes.object,
@@ -219,6 +255,9 @@ HomePage.propTypes = {
 const _getSelectedItem = (items) => {
 	return find(items, {'_selected': true});
 };
+const _getSelectedItems = (items) => {
+	return filter(items, {'_selected': true});
+};
 const mapStateToProps = (state) => {
 	return {
 		barChartData: state.homePage.barChartData,
@@ -226,6 +265,8 @@ const mapStateToProps = (state) => {
 
 		datePickerData: state.homePage.datePicker,
 		shops: state.homePage.shops,
+		shopsDetails: state.homePage.shopsDetails,
+		selectedShops: _getSelectedItems(state.homePage.shops.items),
 		search: state.homePage.search,
 
 		sideMenu: state.sideMenu,
@@ -242,22 +283,28 @@ const mapDispatchToProps = (dispatch) => {
 				type: HOMEPAGE_SET_NEW_DATA,
 			});
 		},
-		setStartDate: (date) => {
+		setStartDate: date => {
 			dispatch({
 				type: HOMEPAGE_SET_START_DATE,
 				payload: date,
 			});
 		},
-		setEndDate: (date) => {
+		setEndDate: date => {
 			dispatch({
 				type: HOMEPAGE_SET_END_DATE,
 				payload: date,
 			});
 		},
-		fetchShops: (params) => {
+		fetchShops: params => {
 			dispatch(fetchShops(params));
 		},
-		setSearchValue: (value) => {
+		setSelectedShops: shopIds => {
+			dispatch(setSelectedShops(shopIds));
+		},
+		fetchShopsDetails: params => {
+			dispatch(fetchShopsDetails(params));
+		},
+		setSearchValue: value => {
 			dispatch(setSearchValue(value));
 		},
 	};